network explorer should just have one topic
